vkd3d: Hoist out some common code for state objects into a new file.
Avoids some obvious code duplication between RTPSO and WGPSO.
